+++ src/Console/ChromeDriverCommand.php
@@ -2,7 +2,11 @@
 
 namespace Laravel\Dusk\Console;
 
+use Exception;
+use GuzzleHttp\Client;
+use GuzzleHttp\Psr7\Utils;
 use Illuminate\Console\Command;
+use Illuminate\Support\Str;
 use Laravel\Dusk\OperatingSystem;
 use Symfony\Component\Process\Process;
 use ZipArchive;
@@ -31,41 +35,7 @@ class ChromeDriverCommand extends Command
     protected $description = 'Install the ChromeDriver binary';
 
     /**
-     * URL to the latest stable release version.
-     *
-     * @var string
-     */
-    protected $latestVersionUrl = 'https://chromedriver.storage.googleapis.com/LATEST_RELEASE';
-
-    /**
-     * URL to the latest release version for a major Chrome version.
-     *
-     * @var string
-     */
-    protected $versionUrl = 'https://chromedriver.storage.googleapis.com/LATEST_RELEASE_%d';
-
-    /**
-     * URL to the ChromeDriver download.
-     *
-     * @var string
-     */
-    protected $downloadUrl = 'https://chromedriver.storage.googleapis.com/%s/chromedriver_%s.zip';
-
-    /**
-     * Download slugs for the available operating systems.
-     *
-     * @var array
-     */
-    protected $slugs = [
-        'linux' => 'linux64',
-        'mac' => 'mac64',
-        'mac-intel' => 'mac64',
-        'mac-arm' => 'mac_arm64',
-        'win' => 'win32',
-    ];
-
-    /**
-     * The legacy versions for the ChromeDriver.
+     * The legacy versions for ChromeDriver.
      *
      * @var array
      */
@@ -107,29 +77,6 @@ class ChromeDriverCommand extends Command
     protected $directory = __DIR__.'/../../bin/';
 
     /**
-     * The default commands to detect the installed Chrome / Chromium version.
-     *
-     * @var array
-     */
-    protected $chromeVersionCommands = [
-        'linux' => [
-            '/usr/bin/google-chrome --version',
-            '/usr/bin/chromium-browser --version',
-            '/usr/bin/chromium --version',
-            '/usr/bin/google-chrome-stable --version',
-        ],
-        'mac-intel' => [
-            '/Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome --version',
-        ],
-        'mac-arm' => [
-            '/Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome --version',
-        ],
-        'win' => [
-            'reg query "HKEY_CURRENT_USER\Software\Google\Chrome\BLBeacon" /v version',
-        ],
-    ];
-
-    /**
      * Execute the console command.
      *
      * @return void
@@ -142,11 +89,11 @@ class ChromeDriverCommand extends Command
 
         $currentOS = OperatingSystem::id();
 
-        foreach ($this->slugs as $os => $slug) {
+        foreach (OperatingSystem::all() as $os) {
             if ($all || ($os === $currentOS)) {
-                $archive = $this->download($version, $slug);
+                $archive = $this->download($version, $os);
 
-                $binary = $this->extract($archive);
+                $binary = $this->extract($version, $archive);
 
                 $this->rename($binary, $os);
             }
@@ -182,11 +129,14 @@ class ChromeDriverCommand extends Command
 
         if ($version < 70) {
             return $this->legacyVersions[$version];
+        } elseif ($version < 115) {
+            return $this->fetchChromeVersionFromUrl($version);
         }
 
-        return trim($this->getUrl(
-            sprintf($this->versionUrl, $version)
-        ));
+        $milestones = $this->resolveChromeVersionsPerMilestone();
+
+        return $milestones['milestones'][$version]['version']
+            ?? throw new Exception('Could not determine the ChromeDriver version.');
     }
 
     /**
@@ -196,22 +146,10 @@ class ChromeDriverCommand extends Command
      */
     protected function latestVersion()
     {
-        $streamOptions = [];
+        $versions = json_decode($this->getUrl('https://googlechromelabs.github.io/chrome-for-testing/last-known-good-versions-with-downloads.json'), true);
 
-        if ($this->option('ssl-no-verify')) {
-            $streamOptions = [
-                'ssl' => [
-                    'verify_peer' => false,
-                    'verify_peer_name' => false,
-                ],
-            ];
-        }
-
-        if ($this->option('proxy')) {
-            $streamOptions['http'] = ['proxy' => $this->option('proxy'), 'request_fulluri' => true];
-        }
-
-        return trim(file_get_contents($this->latestVersionUrl, false, stream_context_create($streamOptions)));
+        return $versions['channels']['Stable']['version']
+            ?? throw new Exception('Could not get the latest ChromeDriver version.');
     }
 
     /**
@@ -222,7 +160,7 @@ class ChromeDriverCommand extends Command
      */
     protected function detectChromeVersion($os)
     {
-        foreach ($this->chromeVersionCommands[$os] as $command) {
+        foreach (OperatingSystem::chromeVersionCommands($os) as $command) {
             $process = Process::fromShellCommandline($command);
 
             $process->run();
@@ -245,17 +183,27 @@ class ChromeDriverCommand extends Command
      * Download the ChromeDriver archive.
      *
      * @param  string  $version
-     * @param  string  $slug
+     * @param  string  $os
      * @return string
      */
-    protected function download($version, $slug)
+    protected function download($version, $os)
     {
-        $url = sprintf($this->downloadUrl, $version, $slug);
+        $url = $this->resolveChromeDriverDownloadUrl($version, $os);
 
-        file_put_contents(
-            $archive = $this->directory.'chromedriver.zip',
-            $this->getUrl($url)
-        );
+        $resource = Utils::tryFopen($archive = $this->directory.'chromedriver.zip', 'w');
+
+        $client = new Client();
+
+        $response = $client->get($url, array_merge([
+            'sink' => $resource,
+            'verify' => $this->option('ssl-no-verify') === false,
+        ], array_filter([
+            'proxy' => $this->option('proxy'),
+        ])));
+
+        if ($response->getStatusCode() < 200 || $response->getStatusCode() > 299) {
+            throw new Exception("Unable to download ChromeDriver from [{$url}].");
+        }
 
         return $archive;
     }
@@ -263,10 +211,11 @@ class ChromeDriverCommand extends Command
     /**
      * Extract the ChromeDriver binary from the archive and delete the archive.
      *
+     * @param  string  $version
      * @param  string  $archive
      * @return string
      */
-    protected function extract($archive)
+    protected function extract($version, $archive)
     {
         $zip = new ZipArchive;
 
@@ -274,7 +223,7 @@ class ChromeDriverCommand extends Command
 
         $zip->extractTo($this->directory);
 
-        $binary = $zip->getNameIndex(0);
+        $binary = $zip->getNameIndex(version_compare($version, '115.0', '<') ? 0 : 1);
 
         $zip->close();
 
@@ -292,7 +241,11 @@ class ChromeDriverCommand extends Command
      */
     protected function rename($binary, $os)
     {
-        $newName = str_replace('chromedriver', 'chromedriver-'.$os, $binary);
+        $binary = str_replace(DIRECTORY_SEPARATOR, '/', $binary);
+
+        $newName = Str::contains($binary, '/')
+            ? Str::after(str_replace('chromedriver', 'chromedriver-'.$os, $binary), '/')
+            : str_replace('chromedriver', 'chromedriver-'.$os, $binary);
 
         rename($this->directory.$binary, $this->directory.$newName);
 
@@ -300,25 +253,77 @@ class ChromeDriverCommand extends Command
     }
 
     /**
+     * Get the Chrome version from URL.
+     *
+     * @return string
+     */
+    protected function fetchChromeVersionFromUrl(int $version)
+    {
+        return trim((string) $this->getUrl(
+            sprintf('https://chromedriver.storage.googleapis.com/LATEST_RELEASE_%d', $version)
+        ));
+    }
+
+    /**
+     * Get the Chrome versions per milestone.
+     *
+     * @return array
+     */
+    protected function resolveChromeVersionsPerMilestone()
+    {
+        return json_decode(
+            $this->getUrl('https://googlechromelabs.github.io/chrome-for-testing/latest-versions-per-milestone-with-downloads.json'), true
+        );
+    }
+
+    /**
+     * Resolve the download URL.
+     *
+     * @return string
+     *
+     * @throws \Exception
+     */
+    protected function resolveChromeDriverDownloadUrl(string $version, string $os)
+    {
+        $slug = OperatingSystem::chromeDriverSlug($os, $version);
+
+        if (version_compare($version, '115.0', '<')) {
+            return sprintf('https://chromedriver.storage.googleapis.com/%s/chromedriver_%s.zip', $version, $slug);
+        }
+
+        $milestone = (int) $version;
+
+        $versions = $this->resolveChromeVersionsPerMilestone();
+
+        /** @var array<string, mixed> $chromedrivers */
+        $chromedrivers = $versions['milestones'][$milestone]['downloads']['chromedriver']
+            ?? throw new Exception('Could not get the ChromeDriver version.');
+
+        return collect($chromedrivers)->firstWhere('platform', $slug)['url']
+            ?? throw new Exception('Could not get the ChromeDriver version.');
+    }
+
+    /**
      * Get the contents of a URL using the 'proxy' and 'ssl-no-verify' command options.
      *
-     * @param  string  $url
-     * @return string|bool
+     * @return string
+     *
+     * @throws Exception
      */
     protected function getUrl(string $url)
     {
-        $contextOptions = [];
+        $client = new Client();
 
-        if ($this->option('proxy')) {
-            $contextOptions['http'] = ['proxy' => $this->option('proxy'), 'request_fulluri' => true];
+        $response = $client->get($url, array_merge([
+            'verify' => $this->option('ssl-no-verify') === false,
+        ], array_filter([
+            'proxy' => $this->option('proxy'),
+        ])));
+
+        if ($response->getStatusCode() < 200 || $response->getStatusCode() > 299) {
+            throw new Exception("Unable to fetch contents from [{$url}].");
         }
 
-        if ($this->option('ssl-no-verify')) {
-            $contextOptions['ssl'] = ['verify_peer' => false];
-        }
-
-        $streamContext = stream_context_create($contextOptions);
-
-        return file_get_contents($url, false, $streamContext);
+        return (string) $response->getBody();
     }
 }