+++ src/Console/DuskCommand.php
@@ -125,11 +125,21 @@ class DuskCommand extends Command
         }
 
         $options = array_values(array_filter($options, function ($option) {
-            return ! Str::startsWith($option, ['--env=', '--pest']);
+            return ! Str::startsWith($option, ['--env=', '--pest', '--ansi', '--no-ansi']);
         }));
 
         if (! file_exists($file = base_path('phpunit.dusk.xml'))) {
             $file = base_path('phpunit.dusk.xml.dist');
+        }
+
+        if (version_compare(Version::id(), '10.0', '>=')) {
+            if ($this->option('ansi')) {
+                $options[] = '--colors=always';
+            }
+
+            if ($this->option('no-ansi')) {
+                $options[] = '--colors=never';
+            }
         }
 
         return array_merge(['-c', $file], $options);
@@ -282,20 +292,6 @@ class DuskCommand extends Command
      */
     protected function refreshEnvironment()
     {
-        // BC fix to support Dotenv ^2.2...
-        if (! method_exists(Dotenv::class, 'create')) {
-            (new Dotenv(base_path()))->overload(); // @phpstan-ignore-line
-
-            return;
-        }
-
-        // BC fix to support Dotenv ^3.0...
-        if (! method_exists(Dotenv::class, 'createMutable')) {
-            Dotenv::create(base_path())->overload();
-
-            return;
-        }
-
         Dotenv::createMutable(base_path())->load();
     }