+++ tests/Unit/ProvidesBrowserTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Laravel\Dusk\Tests;
+namespace Laravel\Dusk\Tests\Unit;
 
 use Laravel\Dusk\Concerns\ProvidesBrowser;
 use Mockery as m;
@@ -23,7 +23,7 @@ class ProvidesBrowserTest extends TestCase
     {
         $browser = m::mock(stdClass::class);
         $browser->shouldReceive('screenshot')->with(
-            'failure-Laravel_Dusk_Tests_ProvidesBrowserTest_test_capture_failures_for-0'
+            'failure-Laravel_Dusk_Tests_Unit_ProvidesBrowserTest_test_capture_failures_for-0'
         );
         $browsers = collect([$browser]);
 
@@ -37,7 +37,18 @@ class ProvidesBrowserTest extends TestCase
     {
         $browser = m::mock(stdClass::class);
         $browser->shouldReceive('storeConsoleLog')->with(
-            'Laravel_Dusk_Tests_ProvidesBrowserTest_test_store_console_logs_for-0'
+            'Laravel_Dusk_Tests_Unit_ProvidesBrowserTest_test_store_console_logs_for-0'
+        );
+        $browsers = collect([$browser]);
+
+        $this->storeConsoleLogsFor($browsers);
+    }
+
+    public function test_truncate_test_name_where_that_name_is_too_long_and_might_cause_issues()
+    {
+        $browser = m::mock(stdClass::class);
+        $browser->shouldReceive('storeConsoleLog')->with(
+            'Laravel_Dusk_Tests_Unit_ProvidesBrowserTest_test_truncate_test_name_where_that_name_is_too_long_and_might_cause_is-0'
         );
         $browsers = collect([$browser]);