+++ tests/Unit/WaitsForElementsTest.php
@@ -1,13 +1,16 @@
 <?php
 
-namespace Laravel\Dusk\Tests;
+namespace Laravel\Dusk\Tests\Unit;
 
 use Facebook\WebDriver\Exception\TimeOutException;
+use Facebook\WebDriver\Remote\RemoteWebElement;
+use Facebook\WebDriver\WebDriver;
+use Facebook\WebDriver\WebDriverWait;
 use Laravel\Dusk\Browser;
+use Laravel\Dusk\ElementResolver;
 use Laravel\Dusk\Tests\Concerns\SwapsUrlGenerator;
 use Mockery as m;
 use PHPUnit\Framework\TestCase;
-use stdClass;
 
 class WaitsForElementsTest extends TestCase
 {
@@ -20,16 +23,20 @@ class WaitsForElementsTest extends TestCase
 
     public function test_when_available()
     {
-        $element = m::mock(stdClass::class);
+        $element = m::mock(RemoteWebElement::class);
         $element->shouldReceive('getText')->andReturn('bar');
         $element->shouldReceive('isDisplayed')->andReturnTrue();
 
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(5, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
         $driver->shouldReceive('findElement')->andReturn($element);
 
-        $resolver = m::mock(stdClass::class);
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('format')->with('foo')->andReturn('body foo');
         $resolver->shouldReceive('findOrFail')->with('foo')->andReturn($element);
+        $resolver->shouldReceive('findOrFail')->with('bar')->andThrow(TimeOutException::class, 'Waited 5 seconds for selector [bar].');
 
         $browser = new Browser($driver, $resolver);
 
@@ -50,7 +57,12 @@ class WaitsForElementsTest extends TestCase
     {
         Browser::$waitSeconds = 2;
 
-        $browser = new Browser(new stdClass);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $browser = new Browser($driver);
         $then = microtime(true);
 
         try {
@@ -68,7 +80,12 @@ class WaitsForElementsTest extends TestCase
     {
         Browser::$waitSeconds = 2;
 
-        $browser = new Browser(new stdClass);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(0, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $browser = new Browser($driver);
         $then = microtime(true);
 
         try {
@@ -84,7 +101,12 @@ class WaitsForElementsTest extends TestCase
 
     public function test_wait_using()
     {
-        $browser = new Browser(new stdClass);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(5, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $browser = new Browser($driver);
 
         $browser->waitUsing(5, 100, function () {
             return true;
@@ -95,7 +117,12 @@ class WaitsForElementsTest extends TestCase
     {
         $this->expectException(TimeOutException::class);
 
-        $browser = new Browser(new stdClass);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(1, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $browser = new Browser($driver);
 
         $browser->waitUsing(1, 100, function () {
             return false;
@@ -104,10 +131,13 @@ class WaitsForElementsTest extends TestCase
 
     public function test_can_wait_for_location()
     {
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
         $driver->shouldReceive('executeScript')
             ->with("return window.location.pathname == '/home';")
             ->andReturnTrue();
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
         $browser = new Browser($driver);
 
@@ -116,10 +146,13 @@ class WaitsForElementsTest extends TestCase
 
     public function test_can_wait_for_a_url_location()
     {
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
         $driver->shouldReceive('executeScript')
             ->with('return `${location.protocol}//${location.host}${location.pathname}` == \'http://example.com/home\';')
             ->andReturnTrue();
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
         $browser = new Browser($driver);
 
@@ -128,10 +161,13 @@ class WaitsForElementsTest extends TestCase
 
     public function test_can_wait_for_a_ssl_url_location()
     {
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
         $driver->shouldReceive('executeScript')
             ->with('return `${location.protocol}//${location.host}${location.pathname}` == \'https://example.com/home\';')
             ->andReturnTrue();
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
         $browser = new Browser($driver);
 
@@ -142,10 +178,13 @@ class WaitsForElementsTest extends TestCase
     {
         $this->swapUrlGenerator();
 
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
         $driver->shouldReceive('executeScript')
             ->with("return window.location.pathname == '/home/';")
             ->andReturnTrue();
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
         $browser = new Browser($driver);
 
@@ -154,56 +193,76 @@ class WaitsForElementsTest extends TestCase
 
     public function test_can_wait_for_text()
     {
-        $element = m::mock(stdClass::class);
+        $element = m::mock(RemoteWebElement::class);
         $element->shouldReceive('getText')->andReturn('Discount: 20%');
 
-        $resolver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('findOrFail')->with('')->andReturn($element);
 
-        $browser = new Browser(new stdClass, $resolver);
+        $browser = new Browser($driver, $resolver);
 
         $browser->waitForText('Discount: 20%');
     }
 
     public function test_can_wait_for_text_to_go_missing()
     {
-        $element = m::mock(stdClass::class);
+        $element = m::mock(RemoteWebElement::class);
         $element->shouldReceive('getText')
             ->times(3)
             ->andReturn('Discount: 20%', 'Discount: 20%', 'SOLD OUT!');
 
-        $resolver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('findOrFail')->with('')->andReturn($element);
 
-        $browser = new Browser(new stdClass, $resolver);
+        $browser = new Browser($driver, $resolver);
 
         $browser->waitUntilMissingText('Discount: 20%');
     }
 
     public function test_wait_until_missing()
     {
-        $element = m::mock(stdClass::class);
+        $element = m::mock(RemoteWebElement::class);
         $element->shouldReceive('isDisplayed')
             ->times(2)
             ->andReturn(true, false);
 
-        $resolver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('findOrFail')->with('foo')->andReturn($element);
 
-        $browser = new Browser(stdClass::class, $resolver);
+        $browser = new Browser($driver, $resolver);
 
         $browser->waitUntilMissing('foo');
     }
 
     public function test_wait_until_missing_text_failure_message_containing_a_percent_character()
     {
-        $element = m::mock(stdClass::class);
+        $element = m::mock(RemoteWebElement::class);
         $element->shouldReceive('getText')->andReturn('Discount: 20%');
 
-        $resolver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(1, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('findOrFail')->with('')->andReturn($element);
 
-        $browser = new Browser(new stdClass, $resolver);
+        $browser = new Browser($driver, $resolver);
 
         try {
             $browser->waitUntilMissingText('Discount: 20%', 1);
@@ -215,13 +274,18 @@ class WaitsForElementsTest extends TestCase
 
     public function test_wait_for_text_failure_message_containing_a_percent_character()
     {
-        $element = m::mock(stdClass::class);
+        $element = m::mock(RemoteWebElement::class);
         $element->shouldReceive('getText')->andReturn('Discount: None');
 
-        $resolver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(1, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('findOrFail')->with('')->andReturn($element);
 
-        $browser = new Browser(new stdClass, $resolver);
+        $browser = new Browser($driver, $resolver);
 
         try {
             $browser->waitForText('Discount: 20%', 1);
@@ -233,14 +297,16 @@ class WaitsForElementsTest extends TestCase
 
     public function test_wait_for_text_in_failure_message_containing_a_percent_character()
     {
-        $element = m::mock(stdClass::class);
-        $element->shouldReceive('getText')->andReturn('Discount: None');
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(1, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
-        $resolver = m::mock(stdClass::class);
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('format')->with('foo')->andReturn('body foo');
-        $resolver->shouldReceive('findOrFail')->with('foo')->andReturn($element);
+        $resolver->shouldReceive('findOrFail')->with('foo')->andThrow(TimeOutException::class, 'Waited 1 seconds for text "Discount: 20%" in selector foo');
 
-        $browser = new Browser(new stdClass, $resolver);
+        $browser = new Browser($driver, $resolver);
 
         try {
             $browser->waitForTextIn('foo', 'Discount: 20%', 1);
@@ -252,9 +318,11 @@ class WaitsForElementsTest extends TestCase
 
     public function test_wait_for_link_failure_message_containing_a_percent_character()
     {
-        $driver = m::mock(stdClass::class);
-        $driver->shouldReceive('executeScript')
-            ->andReturnFalse();
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('executeScript')->andThrow(TimeOutException::class, 'Waited 1 seconds for link [https://laravel.com?q=foo%20bar].');
+        $driver->shouldReceive('wait')->with(1, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
         $browser = new Browser($driver);
 
@@ -268,9 +336,11 @@ class WaitsForElementsTest extends TestCase
 
     public function test_wait_for_location_failure_message_containing_a_percent_character()
     {
-        $driver = m::mock(stdClass::class);
-        $driver->shouldReceive('executeScript')
-            ->andReturnFalse();
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('executeScript')->andThrow(TimeOutException::class, 'Waited 1 seconds for location [https://laravel.com?q=foo%20bar].');
+        $driver->shouldReceive('wait')->with(1, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
         $browser = new Browser($driver);
 
@@ -284,50 +354,68 @@ class WaitsForElementsTest extends TestCase
 
     public function test_wait_for_an_element_to_be_enabled()
     {
-        $element = m::mock(stdClass::class);
+        $element = m::mock(RemoteWebElement::class);
         $element->shouldReceive('isEnabled')->andReturnTrue();
 
-        $resolver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(1, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('findOrFail')->with('#button')->andReturn($element);
 
-        $browser = new Browser(new stdClass, $resolver);
+        $browser = new Browser($driver, $resolver);
 
         $browser->waitUntilEnabled('#button', 1);
     }
 
     public function test_wait_for_an_element_to_be_disabled()
     {
-        $element = m::mock(stdClass::class);
+        $element = m::mock(RemoteWebElement::class);
         $element->shouldReceive('isEnabled')->andReturn(false);
 
-        $resolver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(1, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('findOrFail')->with('#button')->andReturn($element);
 
-        $browser = new Browser(new stdClass, $resolver);
+        $browser = new Browser($driver, $resolver);
 
         $browser->waitUntilDisabled('#button', 1);
     }
 
     public function test_wait_for_text_in()
     {
-        $element = m::mock(stdClass::class);
+        $element = m::mock(RemoteWebElement::class);
         $element->shouldReceive('getText')->andReturn('Discount: 20%');
 
-        $resolver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
+
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('format')->with('foo')->andReturn('body foo');
         $resolver->shouldReceive('findOrFail')->with('foo')->andReturn($element);
 
-        $browser = new Browser(stdClass::class, $resolver);
+        $browser = new Browser($driver, $resolver);
 
         $browser->waitForTextIn('foo', 'Discount: 20%');
     }
 
     public function test_wait_for_link()
     {
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
         $driver->shouldReceive('executeScript')
             ->times(3)
             ->andReturnTrue();
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
         $link = 'https://laravel.com/docs/8.x/dusk';
 
@@ -347,10 +435,13 @@ JS;
 
     public function test_wait_until_vue()
     {
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
         $driver->shouldReceive('executeScript')->andReturn('bar');
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
-        $resolver = m::mock(stdClass::class);
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('format')->with('foo')->andReturn('body foo');
 
         $browser = new Browser($driver, $resolver);
@@ -360,10 +451,13 @@ JS;
 
     public function test_wait_until_vue_is_not()
     {
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
         $driver->shouldReceive('executeScript')->andReturn('bar');
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
-        $resolver = m::mock(stdClass::class);
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('format')->with('foo')->andReturn('body foo');
 
         $browser = new Browser($driver, $resolver);
@@ -373,9 +467,9 @@ JS;
 
     public function test_wait_for_dialog()
     {
-        $driver = m::mock(stdClass::class);
-        $driver->shouldReceive('wait')->andReturn($driver);
-        $driver->shouldReceive('until')->andReturnTrue();
+        $driver = m::mock(WebDriver::class);
+        $driver->shouldReceive('wait')->with(2, 100)->andReturn($driverWait = m::mock(WebDriverWait::class));
+        $driverWait->shouldReceive('until')->andReturnTrue();
 
         $browser = new Browser($driver);
 
@@ -384,10 +478,13 @@ JS;
 
     public function test_wait_for_reload()
     {
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
         $driver->shouldReceive('executeScript')
             ->times(2)
             ->andReturnTrue();
+        $driver->shouldReceive('wait')->with(2, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
         $browser = new Browser($driver);
 
@@ -396,14 +493,17 @@ JS;
 
     public function test_wait_for_event()
     {
-        $driver = m::mock(stdClass::class);
+        $driver = m::mock(WebDriver::class);
         $driver->shouldReceive('manage->timeouts->setScriptTimeout')->with(3);
         $driver->shouldReceive('executeAsyncScript')->with(
             'eval(arguments[0]).addEventListener(arguments[1], () => arguments[2](), { once: true });',
             ['body form', 'submit']
         );
+        $driver->shouldReceive('wait')->with(3, 100)->andReturnUsing(function ($seconds, $interval) use ($driver) {
+            return new WebDriverWait($driver, $seconds, $interval);
+        });
 
-        $resolver = m::mock(stdClass::class);
+        $resolver = m::mock(ElementResolver::class);
         $resolver->shouldReceive('findOrFail')
             ->with('form')
             ->andReturn('body form');